auth-service/backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pika==1.3.1
python-dotenv==1.0.0
pydantic[email]
python-multipart
